requirements.txt
streamlit==1.32.0
tensorflow-intel==2.15.0
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
pillow==10.2.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
h5py==3.10.0
protobuf==3.20.3
typing-extensions==4.9.0
setuptools==69.0.3